+++ db/migrate/20161012151053_rename_dsmb_member_to_unblinded_for_users.rb
@@ -1,0 +1,5 @@
+class RenameDsmbMemberToUnblindedForUsers < ActiveRecord::Migration[5.0]
+  def change
+    rename_column :users, :dsmb_member, :unblinded
+  end
+end