+++ db/migrate/20161012151641_rename_dsmb_only_to_unblinded_only_for_categories.rb
@@ -1,0 +1,5 @@
+class RenameDsmbOnlyToUnblindedOnlyForCategories < ActiveRecord::Migration[5.0]
+  def change
+    rename_column :categories, :dsmb_only, :unblinded_only
+  end
+end